+++ tests/testthat/test-dplyr.R
@@ -4,7 +4,7 @@ context("dplyr verbs for tsibble")
 
 pedestrian <- pedestrian %>%
   group_by(Sensor) %>%
-  slice(1:10) %>%
+  slice(1:48) %>%
   ungroup()
 
 tourism <- tourism %>%
@@ -21,7 +21,7 @@ test_that("group_by()", {
     group_by(Sensor)
   expect_equal(n_groups(grped_df), 4)
   expect_length(group_size(grped_df), 4)
-  expect_length(group_indices(grped_df), 40)
+  expect_length(group_indices(grped_df), NROW(grped_df))
   expect_named(
     pedestrian %>% group_by(sensor = Sensor),
     c(names(pedestrian), "sensor")
@@ -52,9 +52,9 @@ test_that("ungroup()", {
 test_that("arrange.tbl_ts()", {
   expect_equal(arrange(tourism), tourism)
   expect_equal(arrange(tourism %>% group_by(Purpose)), group_by(tourism, Purpose))
-  expect_warning(tsbl1 <- arrange(tourism, Quarter), "Unexpected temporal order.")
+  tsbl1 <- arrange(tourism, Quarter)
   expect_equal(tsbl1, tourism)
-  expect_false(is_ordered(tsbl1))
+  expect_true(is_ordered(tsbl1))
   expect_identical(key(tsbl1), key(tourism))
   expect_identical(groups(tsbl1), groups(tourism))
   tsbl2 <- arrange(tsbl1, Region, State, Purpose, Quarter)
@@ -65,37 +65,31 @@ idx_year <- seq.int(1970, 2010, by = 10)
 dat_x <- tsibble(year = idx_year, value = rnorm(5), index = year)
 
 test_that("warnings for arrange a univariate time series", {
-  expect_warning(arrange(dat_x, value), "Unexpected temporal order.")
+  expect_warning(arrange(dat_x, value), "Unknown temporal order.")
 })
 
 test_that("expect warnings from arrange.tbl_ts()", {
   expect_is(pedestrian %>% arrange(desc(Sensor), Date_Time), "tbl_ts")
-  expect_warning(pedestrian %>% arrange(Time), "Unexpected temporal order.")
-  expect_warning(pedestrian %>% arrange(Sensor, desc(Date_Time)), "Unexpected temporal order.")
-  expect_warning(pedestrian %>% arrange(desc(Date_Time)), "Unexpected temporal order.")
-  expect_warning(pedestrian %>% arrange(Count, Date_Time, Sensor), "Unexpected temporal order.")
-  expect_warning(pedestrian %>% arrange(Sensor, Count, Date_Time), "Unexpected temporal order.")
-  expect_warning(tbl <- pedestrian %>% arrange(Date_Time, Sensor), "Unexpected temporal order.")
+  expect_warning(pedestrian %>% arrange(Time), "Unknown temporal order.")
+  expect_warning(pedestrian %>% arrange(Sensor, desc(Date_Time)), "Unknown temporal order.")
+  expect_warning(pedestrian %>% arrange(desc(Date_Time)), "Unknown temporal order.")
+  expect_warning(pedestrian %>% arrange(Count, Date_Time, Sensor), "Unknown temporal order.")
+  expect_warning(pedestrian %>% arrange(Sensor, Count, Date_Time), "Unknown temporal order.")
+  tbl <- pedestrian %>% arrange(Date_Time, Sensor)
   expect_identical(tbl %>% arrange(Sensor, Date_Time), pedestrian)
   bm <- pedestrian %>%
     filter(Sensor == "Birrarung Marr")
-  expect_warning(bm %>% arrange(desc(Date_Time)), "Unexpected temporal order.")
+  expect_warning(bm %>% arrange(desc(Date_Time)), "Unknown temporal order.")
 })
 
 test_that("arrange.grouped_ts()", {
-  expect_warning(
-    tsbl2 <- tourism %>% group_by(Region, State) %>% arrange(Quarter),
-    "Unexpected temporal order."
-  )
+  tsbl2 <- tourism %>% group_by(Region, State) %>% arrange(Quarter)
   expect_equal(tsbl2, tourism)
   expect_identical(key(tsbl2), key(tourism))
   expect_identical(group_vars(tsbl2), c("Region",  "State"))
-  expect_warning(
-    tsbl3 <- tourism %>%
-      group_by(Region, State) %>%
-      arrange(Quarter, .by_group = TRUE),
-    "Unexpected temporal order."
-  )
+  tsbl3 <- tourism %>%
+    group_by(Region, State) %>%
+    arrange(Quarter, .by_group = TRUE)
   expect_equal(tsbl3, tourism)
   expect_identical(key(tsbl3), key(tourism))
   expect_identical(group_vars(tsbl3), c("Region",  "State"))
@@ -126,7 +120,7 @@ test_that("filter() and slice()", {
   expect_identical(slice(pedestrian, NA), slice(pedestrian, 0L))
   expect_identical(slice(pedestrian, c(1, NA)), slice(pedestrian, 1L))
   expect_identical(slice(pedestrian, c(1, NA, 100000)), slice(pedestrian, 1L))
-  expect_warning(slice(pedestrian, 3:1), "Unexpected temporal order.")
+  expect_warning(slice(pedestrian, 3:1), "Unknown temporal order.")
   expect_error(slice(pedestrian, c(3, 3)), "Duplicated")
   expect_error(slice(pedestrian, 3, 3), "only accepts one expression.")
   expect_identical(slice(pedestrian, -(1:10)), pedestrian[-(1:10), ])